move checking whether a dataset with the import url already exists into import route component
